+++ docs/changelog.rst
@@ -5,6 +5,29 @@ All notable changes to this project will be documented in this file.
 
 The format is based on `Keep a Changelog`_, and this project adheres to `Semantic Versioning`_.
 
+[0.10.3] - XXXX-XX-XX
+---------------------
+
+**Notes**
+
+-
+
+**Added**
+
+-
+
+**Changed**
+
+-
+
+**Fixed**
+
+-
+
+**Removed**
+
+- 
+
 
 [0.10.2] - 2025-06-20
 ---------------------